requirements.txt
auto-tagify==1.4
autopep8==2.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.2.5
Flask-Cors==3.0.10
fsspec==2023.1.0
gunicorn==20.1.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.3.1
python-dotenv==0.21.1
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
